Introduce store package abstractions for writing to persistent storage.
In this commit, we introduce Iterators and develop Appenders.
All Iterator implementations go in iterators.go; all Appender
implementations go in appenders.go to keep store.go smaller.

Also in this commit, we introduce the timestampSeriesType to store
series of just timestamps with no values. Since both timeSeriesType
and timeStampSeriesType are both sequences of pages that store items,
they share common functionality. Therefore we factor out that common
functionality into pageSeriesType. Both timeSeriesType and
timeStampSeriesType have an anonymous pageSeriesType field.

Pages of timestamps and pages of values also share common functionality.
We introduce the basicPageType interface to provide a uniform way to
search a page by timestamp, iterate over a page, clear a page, and get
the number of items in a page.

We add a map of groupId to timeStampSeriesType inside
timeSeriesCollectionType so that we can keep track of all the timestamps
for each endpoint.

To facilitate writing every metric for every timestamp to persistent
storage, we add additional methods to TimeSeriesCollectionType like
FetchForward and FetchForwardWithTimeStamps designed for fetching values
to write to persistent storage.

Finally we introduce the NamedIterator abstraction to encapsulate away
the details of fetching the data to write to persistent storage and
keeping track of what has been written out and what has not. Clients get
a NamedIterator for an endpoint and iterate over it to write the values
out to persistent storage. If writing to persistent storage is
successful, the client can commit its progress.