endpoint vs. single metric going inactive to use same code path.
In this commit, I extract a new function that just handles updating the
individual time series. I use this same function when adding new metrics
and when inactivating an endpoint.

Also in this commit, I fix a bug where the code was not counting a new
timestamp page as a result of metrics or an endpoint going inactive.