Add retry primitive to scotty.
This works like sync.Once except that it allows us to rerun a function
until it succeeds. influx and opentsdb libraries succeed on creation, but
fail on write when actual pstore is down. If kafka is down, the kafka
library fails early on creation instead of on write.

We want the kafka library to fail on write instead of on creation when
the kafka backend is down. kafka can always come back online.
Therefore, we introduce this synchronisation primitive so that we keep
trying to set up the kafka connection when we write until we finally
succeed.