New Gate abstraction.
The purpose of this abstraction is to guarantee that no goroutines are
in a particular critical section of code at a given time. I use this
when processing config file changes.

Rather than adding scores of getter/setter methods to pstore writers to
accomodate config file changes and worrying about all the possible data
races, I just re-create brand new pstore writers from the old ones.
Doing this is tricky because the old writers must be in a consistent
state before they can be used to create new pstore writers. In
particular, the old pstore writers can't be in the middle of writing
data out, and they can't be in the middle of updating the count of
metrics left to write. Failure to ensure this means subtle data races as
I was seeing earlier in my testing.

By using this gate abstraction, I can ensure that the original writers
are in a safe state before copying them.

This package allows one to annotate critical sections of code and
provides methods to drain goroutines from those critical sections by
disabling them. Once a critical section is drained, an additional method
can re-enable the critical section allowing goroutines in once again.

This package provides similar functionality to a RWMutex as obtaining an
exclusive lock on an RWMutex drains goroutines from critical sections of
code that need a read lock. But unlike RWMutex which panics if locked when
already locked, in this package, disabling and enabling critical sections
of code are idempotent operations. Also with this package, one can arrange
for goroutines to skip over disabled critical sections of code rather than
blocking on them. I use that feature when protecting the critical
section that writes to the pstore because I want goroutines responsible
for writing to eventually exit rather than blocking on a forever
disabled critical section of code.