Add store.SetIteratorTo
Also add store.Endpoints as on alternative to writing a Visitor
implementation and use latest timestamp instead of earliest
when writing rolled up values.

store.SetIteratorTo provides a way to position iterator A where iterator B is.

If the user changes the roll up span for a pstore writer at runtime,
simply preserving the metrics from the old pstore writer won't do as
changing the roll up span changes how many metric values must be written
out to the pstore. The only reliable way of adjusting the count of
values left to write when roll up span changes is to re-count them.

We re-count values left to write by positioning the value counting
iterator to where the writing iterator is and zeroing out the count
of values left to write. At the next collection of metrics, the counting
iterator will begin counting metrics left to write from where the
writing iterator is. Eventually, the count of metrics left to write will
be correct.

We only need to use this path when the roll up span for a pstore writer
changes. For all other changes, it is fine to just use the values left
to write from the old pstore writer and to keep the counting iterator at
the same location.