Scotty support config files changing at runtime
Because this PR contains several parts, here is a summary of the
changes.

The critical part of this change is the
pstoreContextType.UpdateFromConfigFile method found around line 611 of
apps/scotty/pstorehandler.go. This function is the main loop that is
executed when the configuration file changes.

The startPStoreLoops function around line 775 was modified to start up
the config file changing loop goroutine. It either reads the config file
from the unix file system or from consul.

Around line 430, the pstoreRunnerType was added to represent the action
of writing to a single pstore.