Introduce keyedqueue package.
A keyedqueue.Queue instance provides communication between the metric
collection goroutines and the goroutine writing to CIS.

A keyedqueue.Queue instance works like a common go channel except rather than
being strictly FIFO, it allows newer CIS data for a particular machine to
overwrite older CIS data for the same machine that is already in the queue.

This difference between a channel and a keyedqueue.Queue is vital because once
an ordinary buffered channel fills up, it drops any new incoming data for a
machine on the floor keeping the stale data for that machine.  On the other
hand, A keyedqueue.Queue instance keeps only the most recent data for each
machine. Even though a keyedqueue.Queue instance is not bounded in size like
a buffered channel, it is bounded by the number of machines in the fleet.

Even if the goroutine writing to CIS falls behind, the keyedqueue.Queue
instance ensures that it still writes the most recent data available for each
machine.