DO NOT MERGE -- Use Async HTTP to talk to CIS.
Code is not tested as CIS is being worked on, but I did use the same CIS
library in a toy program and successfully wrote to CIS earlier today.

Changes in the cis folder are to make the cis library asynchronous;
changes in the scotty folder are to integrate scotty with the new async
CIS library.