Let scotty do the best it can to satisfy queries instead of erroring.
If the time range for a query is so long that the number of rollup
buckets needed exceeds, scotty's limit of 1000, let scotty move the
start time forward to that the number of buckets is 1000 rather than
returning an error. This way, scotty satisfies the query, but may not
return older data. This is fine as proxima doesn't expect scotty to
return all the data every time.