DO NOT MERGE -- Add queuesender.
I implemented the queue that we talked about yesterday.

With this scheme I still only get ~370 requests sent per second or about
22K per minute. I am not using scotty, just a simple stand along program
that calls Send in a loop.

Implementing the queue was rather complicated because of connection
refreshes. Either the sending goroutine or the receiving goroutine can
initiate a connection refresh, but both goroutines have to handle a
connection refresh happening at any time.  Moreover, if a connection goes bad,
both the sending and receiving gorotines may try to refresh the
connection at exactly the same time. In addition, whenever a connection refresh
happens, both the sending and receiving goroutines must receive the new
connection. Finally, the the "next request to be sent" pointer in the
queue must be reset to the front of the queue with each connection refresh.

To facilitate handling connection refreshes, I associate a monotonic
increasing connection ID with each connection. This connection ID  makes
it possible to ensure that each connection gets refreshed only one time
even if both the sending and receiving goroutines initiate a connection
refresh at the same time.

I believe that this implementation is slow because when a connection
does fail, I have to resend everything in the queue with the new
connection I create that didn't get a response with the old connection.

By doing this exercise, I learned that LMM can send unsolicited
responses that don't pair with any request which breaks our assumption
that we can accurately pair responses with requests. LMM will send a
status code 408 if we haven't sent any requests in awhile. Although we
can code for 408, there could be other responses that could come that i
don't know about yet.

Despite my efforts, this implementation won't be as fast as we would
like it to be.

Please feel free to give feedback on the code. There may be a simpler
way to do this than what I found.