Fix bug in scotty paging.
Problem timestampSeriesType.Inactivate() would return a zero timestamp
If a new page is needed to store the inactive timestamp. This in turn
would introduce zero timestamps in the btree thereby corrupting it.

This fix makes timestampSeriesType.Inactivate() return a valid timestamp
even if a new page is needed to store the inactive timestamp.