Add writePositionIndex metric.
The writePositionIndex 0-100 estimates what data scotty is currently writing
to a particular data store. 100 means scotty is writing the most recent
data and is completely caught up; 0 means scotty is writing the earliest
data it has and is way behind. When the index is near 0, data is likely to be
evicted before written.