Fix url forming bug in cloudhealth module.
Upgrading to go1.8 exposed a bug in my code. I was stuffing the entire
URL string in the Path field of net/url. While this wworked fine in earlier
versions of go, it doesn't work in go1.8. To fix, I use url.Parse() instead
of forming the url manually.

I also form the URL in the constructor rather than in the Write method
as forming the URL anew every time in the Write method slows down
writing.