Show cloud health/ cloud watch status in showallapps page.
In this PR, for each machine, scotty shows whether or not it is writing
to cloudhealth and cloudwatch. CH means writing to cloudhealth; CW: 5m means
writing to cloud watch every 5 minutes.

I did some refactoring so that the code that decides whether to write to
cloudwatch or cloudhealth is in one place. Before there was some code
duplication because I had decided to display some cloudwatch information in
scotty as an afterthought.